--- controllers/common.controller.js
@@ -1,1 +1,0 @@
-
